Issue 379: Refactor Template Service to Template Utility.
Renamed the (now) utility functions to have `wvr` prefixed in front of them.
This is not used in core, so relocate it to the utility directory in shared and add appropriate imports.

The `compile()` which is now `wvrCompile()` could potentially be made private and not exposed.
For now, it is just being made available along with the now named `wvrParseProjectedContent()`.

Removed all imports to the now no longer existing template.service.
