+++ projects/wvr-elements/src/lib/shared/utility/index.ts
@@ -1,3 +1,4 @@
 export { obtainConfigPath, weaverBootstrap } from './bootstrap.utility';
 export { debounce } from './decorators.utilty';
+export { wvrCompile, wvrParseProjectedContent } from './template.utility';
 export { wvrTimeout } from './timing.utility';
