+++ projects/wvr-elements/src/public-api.ts
@@ -7,7 +7,6 @@ export * from './lib/core/animation.service';
 export * from './lib/core/component-registry.service';
 export * from './lib/core/effects';
 export * from './lib/core/store';
-export * from './lib/core/template.service';
 export * from './lib/core/wvr-core.module';
 export * from './lib/shared/alignment.enum';
 export * from './lib/shared/config';
